Asked DuetGPT to move getApiKey to separate file
