+++ src/boot.ts
@@ -12,45 +12,16 @@ import path, { dirname } from "path";
 import { BufferMemory } from "langchain/memory";
 import { ChatOpenAI } from "langchain/chat_models/openai";
 import { ConversationChain } from "langchain/chains";
-import { OpenAI } from "langchain/llms/openai";
 import { fileURLToPath } from "url";
+import { getApiKey } from "./utils/apiKey";
 import { getErrorMessage } from "./utils/error";
-import os from "os";
 import packageJson from "../package.json";
 import pc from "picocolors";
 
 // Define file and directory names
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
-const configPath = path.join(os.homedir(), ".duet-gpt");
 const promptPath = path.join(__dirname, "../boot.prompt");
-
-// Function to get API key
-async function getApiKey() {
-  // Check if API key is in environment variables
-  if (process.env.OPEN_AI_KEY) {
-    return process.env.OPEN_AI_KEY;
-  }
-
-  // Check if API key is in config file
-  if (fs.existsSync(configPath)) {
-    const config = JSON.parse(fs.readFileSync(configPath, "utf8"));
-    return config.OPEN_AI_KEY;
-  } else {
-    // Prompt user for API key
-    const apiKey = await p.text({
-      message: "Please enter your OpenAI API key: ",
-      placeholder: "sk-XXX…",
-    });
-    if (p.isCancel(apiKey) || !apiKey) {
-      p.cancel("OpenAI API key is required to run DuetGPT");
-      process.exit(0);
-    }
-    // Save API key to config file
-    fs.writeFileSync(configPath, JSON.stringify({ OPEN_AI_KEY: apiKey }));
-    return apiKey;
-  }
-}
 
 // Main boot function
 export async function boot() {
